Allow watching a file that didn't exist before
If a watch request fails because the file didn't exist, an `OSError` is
correctly raised, but the alias remains in the list of requests.
Therefore it isn't possible to catch the exception, and retry the watch
request once the file has been created.

This patch adds the watch request to the request queue after attempting
to call `_setup_watch`, to ensure it doesn't get left in the queue if
the setup fails.